+++ components/StripeSource.component.d.ts
@@ -1,18 +1,12 @@
-import { EventEmitter, ElementRef } from "@angular/core";
-import { StripeToken, StripeSource as StripeSourceType, StripeInstance } from "../StripeTypes";
+import { EventEmitter } from "@angular/core";
+import { StripeToken, StripeSource as StripeSourceType } from "../StripeTypes";
 import { StripeScriptTag } from "../StripeScriptTag";
-export declare class StripeSource {
-    ElementRef: ElementRef;
+import { StripeComponent } from "./StripeComponent";
+export declare class StripeSource extends StripeComponent {
     StripeScriptTag: StripeScriptTag;
-    catcher: EventEmitter<Error>;
-    invalid: Error;
-    invalidChange: EventEmitter<Error>;
     source: StripeSourceType;
     sourceChange: EventEmitter<StripeSourceType>;
-    stripe: StripeInstance;
     elements: any;
-    constructor(ElementRef: ElementRef, StripeScriptTag: StripeScriptTag);
-    ngOnInit(): void;
-    init(): Promise<StripeInstance>;
-    createSource(extraData?: any): Promise<StripeToken>;
+    constructor(StripeScriptTag: StripeScriptTag);
+    createSource(): Promise<StripeToken>;
 }