+++ components/StripeSource.component.js
@@ -1,4 +1,17 @@
 "use strict";
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
 var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
     var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
     if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
@@ -11,28 +24,21 @@ var __metadata = (this && this.__metadata) || function (k, v) {
 Object.defineProperty(exports, "__esModule", { value: true });
 var core_1 = require("@angular/core");
 var StripeScriptTag_1 = require("../StripeScriptTag");
+var StripeComponent_1 = require("./StripeComponent");
 var stripe_source_pug_1 = require("./templates/stripe-source.pug");
-var StripeSource = /** @class */ (function () {
-    function StripeSource(ElementRef, StripeScriptTag) {
-        this.ElementRef = ElementRef;
-        this.StripeScriptTag = StripeScriptTag;
-        this.catcher = new core_1.EventEmitter();
-        this.invalidChange = new core_1.EventEmitter();
-        this.sourceChange = new core_1.EventEmitter();
+var StripeSource = /** @class */ (function (_super) {
+    __extends(StripeSource, _super);
+    function StripeSource(StripeScriptTag) {
+        var _this = _super.call(this, StripeScriptTag) || this;
+        _this.StripeScriptTag = StripeScriptTag;
+        _this.sourceChange = new core_1.EventEmitter();
+        return _this;
     }
-    StripeSource.prototype.ngOnInit = function () {
-        this.init();
-    };
-    StripeSource.prototype.init = function () {
-        var _this = this;
-        return this.StripeScriptTag.promiseInstance()
-            .then(function (i) { return _this.stripe = i; });
-    };
-    StripeSource.prototype.createSource = function (extraData) {
+    StripeSource.prototype.createSource = function () {
         var _this = this;
         delete this.invalid;
         this.invalidChange.emit(this.invalid);
-        return this.stripe.createSource(this.elements, extraData)
+        return this.stripe.createSource(this.elements)
             .then(function (result) {
             if (result.error) {
                 if (result.error.type == "validation_error") {
@@ -50,18 +56,6 @@ var StripeSource = /** @class */ (function () {
         });
     };
     __decorate([
-        core_1.Output("catch"),
-        __metadata("design:type", core_1.EventEmitter)
-    ], StripeSource.prototype, "catcher", void 0);
-    __decorate([
-        core_1.Input(),
-        __metadata("design:type", Error)
-    ], StripeSource.prototype, "invalid", void 0);
-    __decorate([
-        core_1.Output(),
-        __metadata("design:type", core_1.EventEmitter)
-    ], StripeSource.prototype, "invalidChange", void 0);
-    __decorate([
         core_1.Input(),
         __metadata("design:type", Object)
     ], StripeSource.prototype, "source", void 0);
@@ -75,10 +69,9 @@ var StripeSource = /** @class */ (function () {
             template: stripe_source_pug_1.string,
             exportAs: "StripeSource"
         }),
-        __metadata("design:paramtypes", [core_1.ElementRef,
-            StripeScriptTag_1.StripeScriptTag])
+        __metadata("design:paramtypes", [StripeScriptTag_1.StripeScriptTag])
     ], StripeSource);
     return StripeSource;
-}());
+}(StripeComponent_1.StripeComponent));
 exports.StripeSource = StripeSource;
 //# sourceMappingURL=StripeSource.component.js.map